Update MnM.py
Changed both slider QTE's to work with mouse clicks. Changed the timer slider to change colors to confirm success or failure on click. Changed the slider to change colors as well. Changed the positioning of the sliders so that they appear in the middle of the screen and don't have a background fill.